+++ packages/client/src/store/index.ts
@@ -1,9 +1,8 @@
 import { combineReducers, configureStore } from "@reduxjs/toolkit";
-import userReducer from "./user/reducer";
 import { leaderboardSlice } from "./leaderboard/reducer";
+import userReducer from "./user/reducer";
 
 import notifyReducer from "./notification/reducer";
-import authReducer from "./auth/reducer";
 import { errorSlice } from "./error/reducer";
 
 export const store = configureStore({
@@ -12,7 +11,6 @@ export const store = configureStore({
     leaderboard: leaderboardSlice.reducer,
     user: userReducer,
     notify: notifyReducer,
-    auth: authReducer,
   }),
 });
